refactor: перенос статических переменных для логов и ответов
