+++ common/src/main/java/com/github/jon7even/telegram/constants/DefaultSystemMessagesToSend.java
@@ -11,6 +11,14 @@ import lombok.NoArgsConstructor;
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class DefaultSystemMessagesToSend {
+
+    public static final String ERROR_SEND_TEXT = "что-то произошло с сервисом ответов на сообщения, "
+            + "свяжитесь пожалуйста с администратором";
+
     public static final String WE_NOT_SUPPORT = "мы еще не поддерживаем";
-    public static final String ERROR_TO_EXECUTION_FOR_USER = "Произошла ошибка в процессе ответа пользователю Error:";
+
+    public static final String ERROR_TO_SEND = "Извините, произошла ошибка: ";
+
+    public static final String ERROR_RECEIVE = "что-то произошло с сервисом обработки сообщений, "
+            + "свяжитесь пожалуйста с администратором";
 }
